requirements.txt
flask~=3.0.3
flask-wtf
pillow~=10.3.0
numpy~=1.26.4
wtforms~=3.1.2
werkzeug~=3.0.3